requirements.txt
PyMySQL>=1.1,<2.0
SQLAlchemy>=2.0,<3.0
python-dotenv>=1.0,<2.0
